+++ ET_Project_GUI/ET_Project_GUI/DB_InterFace/MazeORM.cs
@@ -8,7 +8,7 @@ namespace ET_Project_GUI.DB_Interface
     class MazeORM
     {
         public int Collisions { get; set; }
-        public int Difficulty { get; set; }
+        public double Difficulty { get; set; }
         public DateTime Start { get; set; }
         public DateTime Stop  { get; set; }
         public readonly int GameId = 1;
@@ -16,7 +16,7 @@ namespace ET_Project_GUI.DB_Interface
         public MazeORM(IDictionary<string, string> data)
         {
             Collisions = Convert.ToInt32(data["wallBumps"]);
-            Difficulty = Convert.ToInt32(data["difficulty"]);
+            Difficulty = Convert.ToDouble(data["difficulty"]);
             Start = Convert.ToDateTime(data["begin"]);
             Stop = Convert.ToDateTime(data["end"]);
         }
